Use long opt `--no-data` instead of short opt `-d` (#13)
This is a README document, so must be readily comprehensible by its
likely audience. The `-d` short option is not terribly easy to find in
the `mysqldump` man page due to the large number of options starting
with the string `-d`. Therefore, we invoke it as a long option instead.

The -h, -u, and -p short options are likely to be so familiar to MySQL
users as not to warrant expansion to long options for comprehensibility,
so we leave them untouched.